todo tomorrow
I forgot to mention the filtering and the reversing of the urls in the
todo.txt

Probably finished with the minimum required after all of that and the
additions to todo.txt?
Maybe some styling?
Regex restrict the input?
I feel like I’m really starting to know where to look for things. This
is my first time ever working with Ruby on Rails, so far so good.